Extract out security threat, make a common folder for components we might reuse
